### 2017-08-11 #356 @labarba
I'm editing two other new submissions (where contacted reviewers have not responded yet), so if @danielskatz wants to take this one, be my guest.
