### 2017-08-11 #357 @whedon
[REVIEW]: schwimmbad: A uniform interface to parallel processing pools in Python
**Submitting author:** @adrn (<a href="http://orcid.org/0000-0003-0872-7098">Adrian Price-Whelan</a>)
**Repository:** <a href="https://github.com/adrn/schwimmbad" target ="_blank">https://github.com/adrn/schwimmbad</a>
**Version:** v0.2.1
