+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - multiprocessing
   - parallel computing
 authors:
-  - name: Adrian Price-Whelan
+  - name: Adrian M. Price-Whelan
     orcid: 0000-0003-0872-7098
     affiliation: 1
   - name: Daniel Foreman-Mackey
