### 2017-08-28 adrn/schwimmbad#16 @jmjwozniak
Link to Read The Docs broken
Currently points here: http://http//schwimmbad.readthedocs.io/en/stable
