### 2017-08-28 #357 @jmjwozniak
Yes, I was able to run the software today.   I also went through the checkboxes.
