### 2017-08-28 #357 @jmjwozniak
I also filed a minor issue: https://github.com/adrn/schwimmbad/issues/16
