### 2017-08-28 #357 @jmjwozniak
As far as I can tell, the version is not v0.2.1 .  git tag has v0.2.0 , schwimmbad.__version__ has 0.3.dev .
