### 2017-08-28 #357 @jmjwozniak
I do not see notes for "Community guidelines".
