### 2017-08-28 #357 @danielskatz
If there are no performance claims, you can check the box, since it says "Have any performance claims of the software been confirmed?" and there are none.  I will also suggest a change in wording for this item for the future, to make this more clear.
