### 2017-08-30 8b7664f
fixes #16 - update read the docs link
