### 2017-08-30 adrn/schwimmbad#16 @adrn
Thanks @jmjwozniak - fixed!
