### 2017-08-30 #357 @adrn
I was planning to release v0.3 after the comments from this review. The current version is v0.2.1 on pypi (I had forgotten to push that tag, but [fixed now](https://github.com/adrn/schwimmbad/tree/v0.2.1)).

I also fixed the issue with the README.

Thanks!
