### 2017-08-30 #357 @danielskatz
The link in the README points to http://schwimmbad.readthedocs.io/en/stable/, which says "Sorry, This page does not exist yet."

Also, the Community Guidelines issue needs to be addressed...
