### 2017-08-30 a2c6ce1
updated link to main docs page
