### 2017-08-30 4c2cd15
added a page describing how to contribute
