+++ docs/contributing.rst
@@ -1,0 +1,19 @@
+.. _contribute:
+
+**********************************
+How to Contribute or Report Issues
+**********************************
+
+Contributions to the project are always welcome!
+
+If you would like to contribute new features, if it is a significant change to
+the code (left open to interpretation), please start by `opening an issue
+<https://github.com/adrn/schwimmbad/issues>`_ on the GitHub page for this
+project describing your idea. If it's a bug fix or minor change to the source
+code, or a change to the documentation, feel free to instead submit a `pull
+request <https://github.com/adrn/schwimmbad/pulls>`_ on GitHub instead.
+
+If you've found a bug or issue with the code, or something that is unclear in
+the documentation, please `open an issue
+<https://github.com/adrn/schwimmbad/issues>`_ on the GitHub page for this
+project describing the bug or issue.
