### 2017-08-30 #357 @adrn
Strange, looks like the readthedocs "stable" build started failing...Updated the link to the main RTD page instead.

I also added a [How to contribute and report issues](http://schwimmbad.readthedocs.io/en/latest/contributing.html) page to the documentation.
