### 2017-09-06 #357 @arfon
@adrn - Could you please cite the references you have in your `paper.bib` directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)). They're currently not being picked up by Pandoc.
