+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@techreport{Forum:1994,
+@techreport{Forum1994,
   author = {Forum, Message P},
   title = {MPI: A Message-Passing Interface Standard},
   year = {1994},
@@ -7,7 +7,7 @@
   address = {Knoxville, TN, USA},
 }
 
-@InProceedings{Gabriel:2004,
+@InProceedings{Gabriel2004,
   author       = {Edgar Gabriel and Graham E. Fagg and George Bosilca
                   and Thara Angskun and Jack J. Dongarra and Jeffrey
                   M. Squyres and Vishal Sahay and Prabhanjan Kambadur
@@ -24,7 +24,7 @@
   month        = {September}
 }
 
-@article{Dalcin:2005,
+@article{Dalcin2005,
   title = "MPI for Python",
   journal = "Journal of Parallel and Distributed Computing",
   volume = "65",
@@ -42,7 +42,7 @@
   keywords = "Parallel Python"
 }
 
-@article{Dalcin:2008,
+@article{Dalcin2008,
   title = "MPI for Python: Performance improvements and MPI-2 extensions",
   journal = "Journal of Parallel and Distributed Computing",
   volume = "68",
@@ -60,7 +60,7 @@
   keywords = "Parallel Python"
 }
 
-@ARTICLE{Lusk:1996,
+@ARTICLE{Lusk1996,
     author = {Ewing Lusk and Nathan Doss and Anthony Skjellum},
     title = {A high-performance, portable implementation of the MPI message passing interface standard},
     journal = {Parallel Computing},
