### 2017-09-06 96d9ee4
preparing release 0.3
